不使用Service, 而直接在Application Thread中创建ExoPlayer
ExoPlayer的文档建议：

It is recommended that instances are created and accessed from a single
application thread. An application's main thread is ideal. Accessing an
instance from multiple threads is discouraged, however if an
application does wish to do this then it may do so provided that it
ensures accesses are synchronized.